Extract per-format data handlers into engine/src/formatDataHandlers.ts
Moves handlePcdData, handleNpyData, handlePtsData, handleOffData,
handleGltfData, handleXyzVariantData, and handleMtlData out of main.ts.
